# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fiber Banking API
  version: 1.0.0
  description: |
    Dokumentasi OpenAPI untuk API perbankan sederhana berbasis Go Fiber.

    Semua pesan response menggunakan bahasa Indonesia dan format standar:
    - Sukses: `{ data: T, status: true, message: string }`
    - Error: `{ status: false, message: string }`
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/admin/staff:
    get:
      tags: [Admin]
      summary: List staff (admin)
      description: |
        Admin mengambil daftar staff.

        role = `admin`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PerPageQuery"
      responses:
        "200":
          description: Staff berhasil diambil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffListSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      items:
                        - id: cstaff000000000000000001
                          name: Staff Teller
                          email: staff@example.com
                          role: staff
                          created_at: "2026-07-23 15:30:00"
                      pagination:
                        current_page: 1
                        per_page: 10
                        from: 1
                        total: 1
                        last_page: 1
                    status: true
                    message: Staff berhasil diambil
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [Admin]
      summary: Buat staff (admin)
      description: |
        Admin membuat user staff baru.

        role = `admin`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffCreateRequest"
            examples:
              default:
                value:
                  name: Staff Baru
                  email: staffbaru@example.com
                  password: password123
      responses:
        "201":
          description: Staff berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      id: cstaffbaru00000000000001
                      name: Staff Baru
                      email: staffbaru@example.com
                      role: staff
                    status: true
                    message: Staff berhasil dibuat
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/staff/{id}:
    put:
      tags: [Admin]
      summary: Update staff (admin)
      description: |
        Admin memperbarui data staff. Field `password` opsional; jika kosong password tidak diubah.

        role = `admin`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/StaffIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdateRequest"
            examples:
              default:
                value:
                  name: Staff Teller Update
                  email: staff.update@example.com
                  password: passwordbaru123
      responses:
        "200":
          description: Staff berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      id: cstaff000000000000000001
                      name: Staff Teller Update
                      email: staff.update@example.com
                      role: staff
                    status: true
                    message: Staff berhasil diperbarui
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Admin]
      summary: Hapus staff (admin)
      description: |
        Admin menghapus staff secara soft delete.

        role = `admin`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/StaffIdPath"
      responses:
        "200":
          description: Staff berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptySuccessResponse"
              examples:
                success:
                  value:
                    status: true
                    message: Staff berhasil dihapus
        default:
          $ref: "#/components/responses/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Gunakan header `Authorization: Bearer <access_token>`."
  parameters:
    AuthorizationHeader:
      name: Authorization
      in: header
      required: true
      description: Token bearer hasil login.
      schema:
        type: string
        example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
    SearchQuery:
      name: q
      in: query
      required: false
      description: Kata kunci pencarian list.
      schema:
        type: string
        example: customer
    PageQuery:
      name: page
      in: query
      required: false
      description: Nomor halaman. Default `1`.
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    PerPageQuery:
      name: per_page
      in: query
      required: false
      description: Jumlah data per halaman. Default `10`, maksimal `100`.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
        example: 10
...
    StaffIdPath:
      name: id
      in: path
      required: true
      description: ID user staff.
      schema:
        type: string
        example: cstaff000000000000000001
...
  responses:
    ErrorResponse:
      description: Contoh response error standar. Status code aktual dapat berupa 400, 401, 403, 404, 409, 502, atau 500 sesuai kasus.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            error:
              value:
                status: false
                message: Request tidak valid
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: boolean
          example: false
        message:
          type: string
          example: Request tidak valid
    Pagination:
      type: object
      required: [current_page, per_page, from, total, last_page]
      properties:
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        from:
          type: integer
          example: 1
        total:
          type: integer
          format: int64
          example: 25
        last_page:
          type: integer
          example: 3
...
    StaffCreateRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Staff Baru
        email:
          type: string
          format: email
          example: staffbaru@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: password123
    StaffUpdateRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          example: Staff Teller Update
        email:
          type: string
          format: email
          example: staff.update@example.com
        password:
          type: string
          format: password
          nullable: true
          description: Opsional. Jika kosong, password tidak diubah.
          example: passwordbaru123
...
    StaffItem:
      type: object
      required: [id, name, email, role, created_at]
      properties:
        id:
          type: string
          example: cstaff000000000000000001
        name:
          type: string
          example: Staff Teller
        email:
          type: string
          format: email
          example: staff@example.com
        role:
          type: string
          enum: [staff]
          example: staff
        created_at:
          type: string
          example: "2026-07-23 15:30:00"
    StaffData:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
          example: cstaff000000000000000001
        name:
          type: string
          example: Staff Teller
        email:
          type: string
          format: email
          example: staff@example.com
        role:
          type: string
          enum: [staff]
          example: staff
...
    EmptySuccessResponse:
      type: object
      required: [status, message]
      properties:
        data:
          nullable: true
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Berhasil
...
    StaffListSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          type: object
          required: [items, pagination]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/StaffItem"
            pagination:
              $ref: "#/components/schemas/Pagination"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Staff berhasil diambil
    StaffSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          $ref: "#/components/schemas/StaffData"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Staff berhasil dibuat